Add a parcelable pm state saver.
